refactor: relocate `auth.py` for graphql (#4451)
