+++ src/phoenix/server/api/mutations/user_mutations.py
@@ -8,9 +8,9 @@ from strawberry.types import Info
 
 from phoenix.auth import compute_password_hash, validate_email_format, validate_password_format
 from phoenix.db import models
+from phoenix.server.api.auth import HasSecret, IsAdmin, IsAuthenticated, IsNotReadOnly
 from phoenix.server.api.context import Context
 from phoenix.server.api.input_types.UserRoleInput import UserRoleInput
-from phoenix.server.api.mutations.auth import HasSecret, IsAdmin, IsAuthenticated, IsNotReadOnly
 from phoenix.server.api.types.User import User
